==> new scripts/4a_Fit_negative_binomial_models_all_scales.Rmd <==
---
title: "Submodels and final models"
author: "Aidan Brushett and Emerald Arthurs"
date: "`r Sys.Date()`"
output: 
  html_document:
    theme: default
    toc: yes
    toc_float: yes
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      paged.print=TRUE)
```

# 0. Before you begin

Aidan Brushett
M.Sc. Student
University of Victoria    
School of Environmental Studies     
Email: [aidanbrushett@uvic.ca](aidanbrushett@uvic.ca)

Emerald Arthurs
M.Sc. Student
University of Victoria    
School of Environmental Studies     
Email: [emeraldarthurs@uvic.ca](emeraldarthurs@uvic.ca)

***

# 0. Setup

```{r libraries, message=FALSE, warning=FALSE}
rm(list = ls())
#library(MASS)
library(glmmTMB)
#library(lme4)
library(tidyverse)
library(MuMIn)
library(rphylopic)
library(corrr)
library(performance)
#library(PerformanceAnalytics)

#`%nin%` <- Negate(`%in%`)
```

# 1. Import data

We will also apply standardized z-scaling to the data. This is done *per buffer* since technically they are different datasets. 

```{r data for modelling}
covs <- read_csv("./data/processed/OSM_all_covariates_HFI_SBFI_final.csv")

response <- read_csv("./data/processed/OSM_monthly_detections_2021_2022_2023.csv") %>%
  
  filter(species == "red squirrel") %>%
  
  select(-species, -presence) %>%
  
  rename(squirrel = detections)


data <- response %>%
  
  left_join(covs, by = c("array", "site"))

# Make sure there are 20 rows per site/month/year 
data %>% 
  
  group_by(site, month, year) %>%
  
  summarize(n_obs = n()) %>%
  
  arrange(n_obs)
# 20 for everything. Looks good!!

data_scaled <- data %>%
  
  group_by(buffer_dist) %>%
  
  mutate(across(cfi_site:last_col(), ~ as.numeric(scale(.)))) %>%
  
  ungroup(.)

# The mean will be 0 even though we grouped first, since the mean for each buffer is still 0. 
summary(data_scaled)

rm(covs, response)
```

# 2. Determine core model

**NOTE**: Sections 2 through 5 could be fit through a single giant purrr loop containing all 20ish models to save code. Breaking it up this way allowed for better exploration, tweaking, faster runtimes when investigating just comp or just config, and gave some additional insights before fitting the interaction models (section 5). For now will keep it in this format :)

```{r}
names(data)
```
Full model will include: lc_coniferous, lc_mixedwood, lc_broadleaf, lc_shrubs, lc_wetland, lc_wetland_treed, fire_0_15

First, let's try using landcover instead of tree composition first. 

```{r}
core_models <- purrr::map(unique(data$buffer_dist), ~{
  
  cat("\r\r Fitting models at scale:", .x, "meters      ")

  data_subset <- data_scaled %>%
    
    filter(buffer_dist == .x)
  
  m_null <- glmmTMB(squirrel ~ 
                        
                        #lc_coniferous + 
                        #lc_mixedwood + 
                        #lc_broadleaf + 
                        #lc_shrubs +
                        #lc_herbs +
                        #fire_0_15 + 

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  m_fire <- glmmTMB(squirrel ~ 
                        
                        #lc_coniferous + 
                        #lc_mixedwood + 
                        #lc_broadleaf + 
                        #lc_shrubs +
                        #lc_herbs +
                        fire_0_15 + 

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  
  m_forest <- glmmTMB(squirrel ~ 
                        
                        lc_coniferous + 
                        lc_mixedwood + 
                        lc_broadleaf + 
                        #lc_shrubs +
                        #lc_herbs +
                        #fire_0_15 + 

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  m_forest_fire <- glmmTMB(squirrel ~ 
                        
                        lc_coniferous + 
                        lc_mixedwood + 
                        lc_broadleaf +
                        #lc_shrubs +
                        #lc_herbs +
                        fire_0_15 + 

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  
  m_landcover <- glmmTMB(squirrel ~ 
                        
                        lc_coniferous + 
                        lc_mixedwood + 
                        lc_broadleaf + 
                        lc_shrubs +
                        lc_wetland +
                        lc_wetland_treed + 
                        #fire_0_15 + 

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  m_landcover_fire <- glmmTMB(squirrel ~ 
                        
                        lc_coniferous + 
                        lc_mixedwood + 
                        lc_broadleaf + 
                        lc_shrubs +
                        lc_wetland +
                        lc_wetland_treed + 
                        fire_0_15 + 

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  MuMIn::model.sel(m_null, m_fire, m_forest, m_forest_fire, m_landcover, m_landcover_fire) %>%
  
    as.data.frame(.) %>%
    
    select(-`disp((Int))`) %>%
  
    rename_with(~ gsub("cond\\(|disp\\(|\\)|\\(", "", .)) %>%
    
    rename(intercept = `Int`) %>%
    
    mutate(model_name = rownames(.),
           buffer_dist = .x) %>%
    
    rename(delta_subset = delta)

  }
  ) %>%
  
  bind_rows(.) %>%
  
  mutate(delta = AICc - min(AICc)) %>%
  
  arrange(delta)
```

We could also re-produce this process using tree species instead of landcover

```{r}
core_models_spp <- purrr::map(unique(data$buffer_dist), ~{
  
  cat("\r\r Fitting models at scale:", .x, "meters      ")

  data_subset <- data_scaled %>%
    
    filter(buffer_dist == .x)
  
  m_null <- glmmTMB(squirrel ~ 
                        
                        #pct_pinu_ban +
                        #pct_pice_mar +
                        #pct_pice_gla +
                        #pct_popu_tre +
                        #pct_lari_lar +
                        #fire_0_15 +

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  m_fire <- glmmTMB(squirrel ~ 
                        
                        #pct_pinu_ban +
                        #pct_pice_mar +
                        #pct_pice_gla +
                        #pct_popu_tre +
                        #pct_lari_lar +
                        fire_0_15 +

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  
  m_conifspp <- glmmTMB(squirrel ~ 
                        
                        pct_pinu_ban +
                        pct_pice_mar +
                        pct_pice_gla +
                        #pct_popu_tre +
                        #pct_lari_lar +
                        #fire_0_15 +

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  m_conifspp_fire <- glmmTMB(squirrel ~ 
                        
                        pct_pinu_ban +
                        pct_pice_mar +
                        pct_pice_gla +
                        #pct_popu_tre +
                        #pct_lari_lar +
                        fire_0_15 +

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  
  m_allspp <- glmmTMB(squirrel ~ 
                        
                        pct_pinu_ban +
                        pct_pice_mar +
                        pct_pice_gla +
                        pct_popu_tre +
                        pct_lari_lar +
                        #fire_0_15 +

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  m_allspp_fire <- glmmTMB(squirrel ~ 
                        
                        pct_pinu_ban +
                        pct_pice_mar +
                        pct_pice_gla +
                        pct_popu_tre +
                        pct_lari_lar +
                        fire_0_15 +

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  MuMIn::model.sel(m_null, m_fire, m_conifspp, m_conifspp_fire, m_allspp, m_allspp_fire) %>%
  
    as.data.frame(.) %>%
    
    select(-`disp((Int))`) %>%
  
    rename_with(~ gsub("cond\\(|disp\\(|\\)|\\(", "", .)) %>%
    
    rename(intercept = `Int`) %>%
    
    mutate(model_name = rownames(.),
           buffer_dist = .x) %>%
    
    rename(delta_subset = delta)

  }
  ) %>%
  
  bind_rows(.) %>%
  
  mutate(delta = AICc - min(AICc)) %>%
  
  arrange(delta)
```

What are the top predictors and top spatial scale for our natural covariates?

```{r}
core_predictors <- core_models %>%
  
  select(1:which(names(.) == "df") - 1, -intercept) %>% 
  
  filter(row_number()==1) %>%
  
  select(where(~ !is.na(.))) %>%  # Remove columns with NA values

  colnames(.)
  
core_buffer <- core_models$buffer_dist[1]

core_predictors
core_buffer
```

# 3. Composition models

```{r}
comp_models <- purrr::map(unique(data$buffer_dist), ~{
  
  cat("\r\r Fitting models at scale:", .x, "meters      ")

  data_subset <- data_scaled %>%
    
    filter(buffer_dist == .x)
  
  m_null <- glmmTMB(squirrel ~ 
                        #lc_coniferous + 
                        #lc_mixedwood + 
                        #lc_broadleaf + 
                        #fire_0_15 + 

                        #pipe_trans + 
                        #seismic + 
                        #wells_inactive + 
                        #harvest_total +
                        #wells_active +
                        #osm_industrial +
                        #harvest_0_15
                        #roads +
                        
                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  m_core <- glmmTMB(squirrel ~ 
                        lc_coniferous + 
                        lc_mixedwood + 
                        lc_broadleaf + 
                        fire_0_15 + 

                        #pipe_trans + 
                        #seismic + 
                        #wells_inactive + 
                        #harvest_total +
                        #wells_active +
                        #osm_industrial +
                        #harvest_0_15
                        #roads +

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  m_seis_wells <- glmmTMB(squirrel ~ 
                        lc_coniferous + 
                        lc_mixedwood + 
                        lc_broadleaf + 
                        fire_0_15 + 
                          
                        pipe_trans + 
                        seismic + 
                        wells_inactive + 
                        #harvest_total +
                        #wells_active +
                        #osm_industrial +
                        #harvest_0_15 +
                        #roads + 

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  m_harvest <- glmmTMB(squirrel ~ 
                        lc_coniferous + 
                        lc_mixedwood + 
                        lc_broadleaf + 
                        fire_0_15 + 
                         
                        #pipe_trans + 
                        #seismic + 
                        #wells_inactive + 
                        #harvest_total +
                        #wells_active +
                        #osm_industrial +
                        harvest_0_15 +
                        #roads + + 

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  m_harvest_seis <- glmmTMB(squirrel ~ 
                        lc_coniferous + 
                        lc_mixedwood + 
                        lc_broadleaf + 
                        fire_0_15 + 
                              
                        pipe_trans + 
                        seismic + 
                        wells_inactive + 
                        #harvest_total +
                        #wells_active +
                        #osm_industrial +
                        harvest_0_15 +
                        #roads + 

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  
  m_active <- glmmTMB(squirrel ~ 
                        lc_coniferous + 
                        lc_mixedwood + 
                        lc_broadleaf + 
                        fire_0_15 + 
                        
                        #pipe_trans + 
                        #seismic + 
                        #wells_inactive + 
                        #harvest_total +
                        wells_active +
                        osm_industrial +
                        #harvest_0_15 +
                        #roads + 

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  m_active_harvest <- glmmTMB(squirrel ~ 
                        lc_coniferous + 
                        lc_mixedwood + 
                        lc_broadleaf + 
                        fire_0_15 + 
                          
                        #pipe_trans + 
                        #seismic + 
                        #wells_inactive + 
                        #harvest_total +
                        wells_active +
                        osm_industrial +
                        harvest_0_15 +
                        #roads + 

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  m_active_seis <- glmmTMB(squirrel ~ 
                        lc_coniferous + 
                        lc_mixedwood + 
                        lc_broadleaf + 
                        fire_0_15 + 
                        
                        pipe_trans + 
                        seismic + 
                        wells_inactive + 
                        #harvest_total +
                        wells_active +
                        osm_industrial +
                        #harvest_0_15 +
                        #roads + 
                          
                      (1|array/site),
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)

  m_global <- glmmTMB(squirrel ~ 
                        lc_coniferous + 
                        lc_mixedwood + 
                        lc_broadleaf + 
                        fire_0_15 + 
                        
                        pipe_trans + 
                        seismic + 
                        wells_inactive + 
                        #harvest_total +
                        wells_active +
                        osm_industrial +
                        harvest_0_15 +
                        #roads + 

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  MuMIn::model.sel(m_null, m_core, m_seis_wells, m_harvest, m_harvest_seis, m_active_seis, m_active, m_active_harvest, m_global) %>%
  
    as.data.frame(.) %>%
    
    select(-`disp((Int))`) %>%
  
    rename_with(~ gsub("cond\\(|disp\\(|\\)|\\(", "", .)) %>%
    
    rename(intercept = `Int`) %>%
    
    mutate(model_name = rownames(.),
           buffer_dist = .x) %>%
    
    rename(delta_subset = delta)

  }
  ) %>%
  
  bind_rows(.) %>%
  
  mutate(delta = AICc - min(AICc)) %>%
  
  arrange(delta)
```

What are the top predictors and top spatial scale for our natural covariates?

```{r}
comp_predictors <- comp_models %>%
  
  select(1:which(names(.) == "df") - 1, -intercept) %>% 
  
  filter(row_number()==1) %>%
  
  select(where(~ !is.na(.))) %>%  # Remove columns with NA values

  colnames(.)
  
comp_buffer <- comp_models$buffer_dist[1]

comp_predictors
comp_buffer
```


# 4. Configuration models

Let's fit the configuration models

```{r}
config_models <- purrr::map(unique(data$buffer_dist), ~{
  
  cat("\r\r Fitting models at scale:", .x, "meters      ")

  data_subset <- data_scaled %>%
    
    filter(buffer_dist == .x)
  
  m_null <- glmmTMB(squirrel ~ 
                        
                        # natural covariates
                        #fire_0_15 +
                        #lc_broadleaf +
                        #lc_coniferous +
                        #lc_mixedwood + 
                      
                        # configuration variables
                        #landscape_shei +
                        #nonanthro_ed + 
                        #landscape_mesh +

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  m_core <- glmmTMB(squirrel ~ 
                        
                        # natural covariates
                        fire_0_15 +
                        lc_broadleaf +
                        lc_coniferous +
                        lc_mixedwood + 
                      
                        # configuration variables
                        #landscape_shei +
                        #nonanthro_ed + 
                        #landscape_mesh +

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  
  m_het <- glmmTMB(squirrel ~ 
                        
                        # natural covariates
                        fire_0_15 +
                        lc_broadleaf +
                        lc_coniferous +
                        lc_mixedwood + 
                      
                        # configuration variables
                        landscape_shei +
                        #nonanthro_ed + 
                        #landscape_mesh +
                         
                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  m_edge <- glmmTMB(squirrel ~ 
                        
                        # natural covariates
                        fire_0_15 +
                        lc_broadleaf +
                        lc_coniferous +
                        lc_mixedwood + 
                      
                        # configuration variables
                        #landscape_shei +
                        nonanthro_ed + 
                        #landscape_mesh +

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
                    
  m_conn <- glmmTMB(squirrel ~ 
                        
                        # natural covariates
                        fire_0_15 +
                        lc_broadleaf +
                        lc_coniferous +
                        lc_mixedwood + 
                      
                        # configuration variables
                        #landscape_shei +
                        #nonanthro_ed + 
                        landscape_mesh +

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  m_edge_conn <- glmmTMB(squirrel ~ 
                        
                        # natural covariates
                        fire_0_15 +
                        lc_broadleaf +
                        lc_coniferous +
                        lc_mixedwood +
                      
                        # configuration variables
                        #landscape_shei +
                        nonanthro_ed + 
                        landscape_mesh +

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)

  m_het_conn <- glmmTMB(squirrel ~ 
                        
                        # natural covariates
                        fire_0_15 +
                        lc_broadleaf +
                        lc_coniferous +
                        lc_mixedwood + 
                      
                        # configuration variables
                        landscape_shei +
                        #nonanthro_ed + 
                        landscape_mesh +

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
    
  m_het_edge <- glmmTMB(squirrel ~
                        
                        # natural covariates
                        fire_0_15 +
                        lc_broadleaf +
                        lc_coniferous +
                        lc_mixedwood +
                      
                        # configuration variables
                        landscape_shei +
                        nonanthro_ed +
                        #landscape_mesh +

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  m_config <- glmmTMB(squirrel ~ 
                        
                        # natural covariates
                        fire_0_15 +
                        lc_broadleaf +
                        lc_coniferous +
                        lc_mixedwood +
                      
                        # configuration variables
                        landscape_shei +
                        nonanthro_ed +
                        landscape_mesh +

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)

  
  MuMIn::model.sel(m_null, m_core, m_edge, m_het, m_conn, m_het_edge, m_edge_conn, m_het_conn, m_config) %>%
  
    as.data.frame(.) %>%
    
    select(-`disp((Int))`) %>%
  
    rename_with(~ gsub("cond\\(|disp\\(|\\)|\\(", "", .)) %>%
    
    rename(intercept = `Int`) %>%
    
    mutate(model_name = rownames(.),
           buffer_dist = .x) %>%
    
    rename(delta_subset = delta)

  }
  ) %>%
  
  bind_rows(.) %>%
  
  mutate(delta = AICc - min(AICc)) %>%
  
  arrange(delta)
```

What are the top predictors and top spatial scale for our configuration covariates?

```{r}
config_predictors <- config_models %>%
  
  select(1:which(names(.) == "df") - 1, -intercept) %>% 
  
  filter(row_number()==1) %>%
  
  select(where(~ !is.na(.))) %>%  # Remove columns with NA values

  colnames(.)
  
config_buffer <- config_models$buffer_dist[1]

config_predictors
config_buffer
```

# 5. Configuration interaction models

```{r}
config_int_models <- purrr::map(unique(data$buffer_dist), ~{
  
  cat("\r\r Fitting models at scale:", .x, "meters      ")

  data_subset <- data_scaled %>%
    
    filter(buffer_dist == .x)
  
  m_null <- glmmTMB(squirrel ~ 
                        
                        # natural covariates
                        #fire_0_15 +
                        #lc_broadleaf +
                        #lc_coniferous +
                        #lc_mixedwood + 
                      
                        # configuration variables
                        #landscape_shei +
                        #nonanthro_ed + 
                        #landscape_mesh +

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  m_core <- glmmTMB(squirrel ~ 
                        
                        # natural covariates
                        fire_0_15 +
                        lc_broadleaf +
                        lc_coniferous +
                        lc_mixedwood + 
                      
                        # configuration variables
                        #landscape_shei +
                        #nonanthro_ed + 
                        #landscape_mesh +

                        (1|array/site),
                      
                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  m_edge <- glmmTMB(squirrel ~ 
                        
                        # natural covariates
                        fire_0_15 +
                        lc_broadleaf +
                        lc_coniferous +
                        lc_mixedwood + 
                      
                        # configuration variables
                        #landscape_shei +
                        nonanthro_ed + 
                        #landscape_mesh +

                        (1|array/site),

                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  m_edge_cfi <- glmmTMB(squirrel ~ 
                        
                        # natural covariates
                        fire_0_15 +
                        lc_broadleaf +
                        lc_coniferous +
                        lc_mixedwood + 
                      
                        # configuration variables
                        #landscape_shei +
                        nonanthro_ed + 
                        cfi_site +
                        #landscape_mesh +

                        (1|array/site),

                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)
  
  m_edgeXcfi <- glmmTMB(squirrel ~ 
                        
                        # natural covariates
                        fire_0_15 +
                        lc_broadleaf +
                        lc_coniferous +
                        lc_mixedwood + 
                      
                        # configuration variables
                        #landscape_shei +
                        nonanthro_ed + 
                        cfi_site +
                        nonanthro_ed*cfi_site +
                        #landscape_mesh +

                        (1|array/site),

                      data = data_subset, 
                      family = nbinom2, 
                      na.action = na.fail)

  MuMIn::model.sel(m_null, m_core, m_edge, m_edge_cfi, m_edgeXcfi) %>%
  
    as.data.frame(.) %>%
    
    select(-`disp((Int))`) %>%
  
    rename_with(~ gsub("cond\\(|disp\\(|\\)|\\(", "", .)) %>%
    
    rename(intercept = `Int`) %>%
    
    mutate(model_name = rownames(.),
           buffer_dist = .x) %>%
    
    rename(delta_subset = delta)

  }
  ) %>%
  
  bind_rows(.) %>%
  
  mutate(delta = AICc - min(AICc)) %>%
  
  arrange(delta)
```

# 6. Combine all models into a bigass table

```{r}
all_models <- bind_rows(
  core_models,
  comp_models,
  config_models,
  config_int_models
  ) %>%
  
  mutate(model_name = paste0(model_name, "_", buffer_dist)) %>%
  
  distinct(model_name, .keep_all = TRUE) %>%
  
  mutate(delta = AICc - min(AICc)) %>%

  relocate(model_name, 'scale' = 'buffer_dist', 
           df, logLik, AICc, delta,
           'delta_subsetted' = 'delta_subset', 
           'weight_subsetted' = 'weight')
  

write_csv(all_models, file = "./data/processed/all_models_scales_aic_with_coefficients.csv")
```


```{r}
all_models <- read_csv("./data/processed/all_models_scales_aic_with_coefficients.csv")
```

```{r}
top_models_scale <- all_models %>%
  
  group_by(scale) %>%
  
  mutate(delta_scale = AICc - min(AICc)) %>%
  
  filter(delta_scale <= 2) %>%
  
  filter(!grepl("core", model_name)) %>%
  
  select(-contains("subsetted"),
         -delta) %>%
  
  arrange(scale, AICc) %>%
  
  relocate(delta_scale, .after = AICc)
```

```{r}
top_models_scale %>%
  
  select(scale, intercept:last_col()) %>%
  
  pivot_longer(., cols = intercept:last_col()) %>%
  
  drop_na(.) %>%
  
  group_by(scale, name) %>%
  
  summarize(value = mean(value)) %>%
  
  filter(name != "intercept") %>%
  
  # Adding our non-250m increment scales messes up the figure. We're fine this way I think for now....
  filter(scale >= 250) %>%
  
  ggplot(., aes(x = scale)) +
  
    geom_tile(aes(y = name, fill = value), color = "grey40") + 
  
    labs(x = "Spatial scale (m)",
         y = "",
         # Not realllyyyy model averaged since it didn't get adjusted using AIC weights 
         color = "Model-\naveraged \nCoefficient",
         fill = "Model-\naveraged \ncoefficient",
         size = "Model-\naveraged \nCoefficient",
         title = "Predictors included in any top models (ΔAICc < 2)") +
  
    scale_x_continuous(expand = c(0,0)) +
  
    scale_fill_gradient2(
      low = "blue3",
      mid = "white",
      high = "firebrick",
      midpoint = 0)  +

    theme_bw() + 
  
    theme(
      panel.background = element_rect(fill = "grey70"),
      panel.grid = element_blank()  # optional: removes gridlines
    )

ggsave("./figures/top_model_scale_analysis_neg_binomial.png", width = 9, height = 4.2)
```
